imap: move over MailboxAttr
